Store actual MySQL rows as arrays in memcache
So apparently memcache was always missing, since it was only ever
storing the result identifier in the cache, which is useless crap. Now
the mysqlQuery method will return an array of all the data returned by
the query (as arrays with stripslashes mapped) using MYSQL_BOTH, or a
boolean when appropriate, and all its dependent methods are updated
accordingly.